mvvm architecture and creating DogRepository.kt
