+++ app/src/main/java/com/navi/dogbreedapp/doglist/DogListActivity.kt
@@ -1,12 +1,15 @@
  package com.navi.dogbreedapp.doglist
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.navi.dogbreedapp.DogModel
 import com.navi.dogbreedapp.databinding.ActivityDogListBinding
 
  class DogListActivity : AppCompatActivity() {
+
+     private val dogListViewModel: DogListViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val binding = ActivityDogListBinding.inflate(layoutInflater)
@@ -17,53 +20,8 @@ import com.navi.dogbreedapp.databinding.ActivityDogListBinding
         val dogAdapter = DogAdapter()
 
         recycler.adapter = dogAdapter
-        dogAdapter.submitList(getFakeDogs())
-    }
-
-    private fun getFakeDogs(): MutableList<DogModel> {
-        val dogList = mutableListOf<DogModel>()
-        dogList.add(
-            DogModel(
-                1, 1, "Chihuahua", "Toy", 5.4,
-                6.7, "", "12 - 15", "", 10.5,
-                12.3
-            )
-        )
-        dogList.add(
-            DogModel(
-                2, 1, "Labrador", "Toy", 5.4,
-                6.7, "", "12 - 15", "", 10.5,
-                12.3
-            )
-        )
-        dogList.add(
-            DogModel(
-                3, 1, "Retriever", "Toy", 5.4,
-                6.7, "", "12 - 15", "", 10.5,
-                12.3
-            )
-        )
-        dogList.add(
-            DogModel(
-                4, 1, "San Bernardo", "Toy", 5.4,
-                6.7, "", "12 - 15", "", 10.5,
-                12.3
-            )
-        )
-        dogList.add(
-            DogModel(
-                5, 1, "Husky", "Toy", 5.4,
-                6.7, "", "12 - 15", "", 10.5,
-                12.3
-            )
-        )
-        dogList.add(
-            DogModel(
-                6, 1, "Xoloscuincle", "Toy", 5.4,
-                6.7, "", "12 - 15", "", 10.5,
-                12.3
-            )
-        )
-        return dogList
+        dogListViewModel.dogList.observe(this) {dogList ->
+            dogAdapter.submitList(dogList)
+        }
     }
 }
