+++ app/src/main/java/com/navi/dogbreedapp/doglist/DogListViewModel.kt
@@ -1,0 +1,26 @@
+package com.navi.dogbreedapp.doglist
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.navi.dogbreedapp.DogModel
+import kotlinx.coroutines.launch
+
+class DogListViewModel: ViewModel() {
+
+    private val _dogList = MutableLiveData<List<DogModel>>()
+    val dogList: LiveData<List<DogModel>> get() = _dogList
+
+    private val dogRepository = DogRepository()
+
+    init {
+        downloadDogs()
+    }
+
+    private fun downloadDogs() {
+        viewModelScope.launch {
+            _dogList.value = dogRepository.downloadDogs()
+        }
+    }
+}
