+++ app/src/main/java/com/navi/dogbreedapp/doglist/DogRepository.kt
@@ -1,0 +1,60 @@
+package com.navi.dogbreedapp.doglist
+
+import com.navi.dogbreedapp.DogModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+class DogRepository {
+    suspend fun downloadDogs(): List<DogModel> {
+        return withContext(Dispatchers.IO) {
+            getFakeDogs()
+        }
+    }
+
+    private fun getFakeDogs(): MutableList<DogModel> {
+        val dogList = mutableListOf<DogModel>()
+        dogList.add(
+            DogModel(
+                1, 1, "Chihuahua", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            DogModel(
+                2, 1, "Labrador", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            DogModel(
+                3, 1, "Retriever", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            DogModel(
+                4, 1, "San Bernardo", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            DogModel(
+                5, 1, "Husky", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            DogModel(
+                6, 1, "Xoloscuincle", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        return dogList
+    }
+}
